Introduce immutable/mutable Graphs
