+++ networkit/cpp/graph/RandomMaximumSpanningForest.cpp
@@ -106,7 +106,7 @@ std::vector<bool> RandomMaximumSpanningForest::getAttribute(bool move) {
 }
 
 Graph RandomMaximumSpanningForest::getMSF(bool move) {
-    Graph result;
+    GraphW result;
 
     if (!hasMSF)
         throw std::runtime_error("Error: The run() method must be executed first");
